move the load function to the test utilities
